+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- concord-agent: normalize dependency URLs, support for Nexus/WARM
+redirects;
 - concord-runner: fix the issue with tasks based on `concord-sdk`;
 - concord-project-model: fix serialization issues when `set` task
 used with nested structures.